+++ src/Action.php
@@ -22,6 +22,7 @@ class Action
         'unset',
         'reset',
         'handle',
+        'cron',
     ];
 
     /**